+++ src/GUI/gui.py
@@ -45,10 +45,13 @@ class MainView(QMainWindow):
                 "editor_font_size": "20",
                 "profile_name": "user",
                 "theme": "dark",
-                "version": "",
                 "session": {
                     "opened_folder": "",
                     "opened_file": ""
+                },
+                "startup": {
+                    "pre_template": "1",
+                    "template": ""
                 }
             }
             with open('config.json', 'w') as file:
@@ -317,8 +320,8 @@ class MainView(QMainWindow):
         paint_action.triggered.connect(self.paint_core)
         utility_menu.addAction(paint_action)
         utility_menu.addSeparator()
-        pbinfo_tools_action = QAction("Pbinfo submit code", self)
-        pbinfo_tools_action.triggered.connect(self.pbinfo_tools_core)
+        pbinfo_tools_action = QAction("Quick submit", self)
+        #pbinfo_tools_action.triggered.connect(self.pbinfo_tools_core)
         utility_menu.addAction(pbinfo_tools_action)
         kilo_tools_action = QAction("Kilonova tools", self)
         kilo_tools_action.triggered.connect(self.kilo_tools_core)
@@ -333,6 +336,9 @@ class MainView(QMainWindow):
         menubar.addMenu(textures_menu)
         menubar.addMenu(utility_menu)
 
+    def re_zoom(self, val):
+        self.config["editor_font_size"] = val
+        self.editor.apply_settings()
     def update_core(self):
         internal.check_for_updates(self.status_bar)
 
@@ -349,9 +355,11 @@ class MainView(QMainWindow):
     def reset_session_core(self):
         session.reset_session()
 
+    """
     def pbinfo_tools_core(self):
         self.pbinfo_win = pbinfo.PbinfoInterface(self.editor, self.theme, self.config,parent=self)
         self.pbinfo_win.show()
+    """
 
     def kilo_tools_core(self):
         self.kilo_win = kilo_tools.Kilotools(self.theme,parent=self)