+++ CHANGELOG.md
@@ -1,4 +1,19 @@
-# Version 2.1.5
+# Version 2.5.0
+
+### Added:
+
+1. New theme light-fire
+2. Added a new menu in right panel "Submit code" that will host the submit function on platforms
+3. Extended the functionality of the Testing tab by adding up to 5 test sets
+4. Settings tab in right panel
+
+### Fixed/Changes:
+
+1. New highlighter color for: light, dark, ocean, dark-blue themes
+2. Fixed autocompletion templates in editor
+3. Fixed documentation bug
+
+# Version 2.2
 
 ### Added: